third_party/WebKit/LayoutTests/inspector/console/console-format-expected.txt

CONSOLE MESSAGE: line 26: test,test2,,,test4,,,,,
CONSOLE MESSAGE: line 27: %o
CONSOLE MESSAGE: line 28: %O
CONSOLE MESSAGE: line 29: Test for zero "%f" in formatter
CONSOLE MESSAGE: line 30: %% self-escape1
CONSOLE MESSAGE: line 31: %%s self-escape2
CONSOLE MESSAGE: line 32: %%ss self-escape3
CONSOLE MESSAGE: line 33: %%s%s%%s self-escape4
CONSOLE MESSAGE: line 34: %%%%% self-escape5
CONSOLE MESSAGE: line 35: %%%s self-escape6
CONSOLE MESSAGE: line 12: /^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
CONSOLE MESSAGE: line 13: /^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
CONSOLE MESSAGE: line 12: /foo\\bar\sbaz/i
CONSOLE MESSAGE: line 13: /foo\\bar\sbaz/i
CONSOLE MESSAGE: line 12: test
CONSOLE MESSAGE: line 13: test
CONSOLE MESSAGE: line 12: test named "test"
CONSOLE MESSAGE: line 13: test named "test"
CONSOLE MESSAGE: line 12: Error
CONSOLE MESSAGE: line 13: Error
CONSOLE MESSAGE: line 12: Error: my error message
CONSOLE MESSAGE: line 13: Error: my error message
CONSOLE MESSAGE: line 12: Error: my multiline
error message
CONSOLE MESSAGE: line 13: Error: my multiline
error message
CONSOLE MESSAGE: line 12: [object HTMLParagraphElement]
CONSOLE MESSAGE: line 13: [object HTMLParagraphElement]
CONSOLE MESSAGE: line 12: function () { return 1; }
CONSOLE MESSAGE: line 13: function () { return 1; }
CONSOLE MESSAGE: line 12: function () {
        return 2;
    }
CONSOLE MESSAGE: line 13: function () {
        return 2;
    }
CONSOLE MESSAGE: line 12: 0.12
CONSOLE MESSAGE: line 13: 0.12
CONSOLE MESSAGE: line 12: http://webkit.org/
CONSOLE MESSAGE: line 13: http://webkit.org/
CONSOLE MESSAGE: line 12: null
CONSOLE MESSAGE: line 13: 
CONSOLE MESSAGE: line 12: undefined
CONSOLE MESSAGE: line 13: 
CONSOLE MESSAGE: line 12: [object Attr]
CONSOLE MESSAGE: line 13: [object Attr]
CONSOLE MESSAGE: line 12: [object Attr]
CONSOLE MESSAGE: line 13: [object Attr]
CONSOLE MESSAGE: line 12: [object Attr]
CONSOLE MESSAGE: line 13: [object Attr]
CONSOLE MESSAGE: line 12: [object Object]
CONSOLE MESSAGE: line 13: [object Object]
CONSOLE MESSAGE: line 12: NaN
CONSOLE MESSAGE: line 13: NaN
CONSOLE MESSAGE: line 12: Infinity
CONSOLE MESSAGE: line 13: Infinity
CONSOLE MESSAGE: line 12: -Infinity
CONSOLE MESSAGE: line 13: -Infinity
CONSOLE MESSAGE: line 12: test,test2,,,test4,,,,,
CONSOLE MESSAGE: line 13: test,test2,,,test4,,,,,
CONSOLE MESSAGE: line 12: [object Object]
CONSOLE MESSAGE: line 13: [object Object]
CONSOLE MESSAGE: line 12: function () {}
CONSOLE MESSAGE: line 13: function () {}
CONSOLE MESSAGE: line 12: [object Object]
CONSOLE MESSAGE: line 13: [object Object]
CONSOLE MESSAGE: line 12: [object SVGSVGElement]
CONSOLE MESSAGE: line 13: [object SVGSVGElement]
CONSOLE MESSAGE: line 12: [object Object]
CONSOLE MESSAGE: line 13: [object Object]
CONSOLE MESSAGE: line 12: 0
CONSOLE MESSAGE: line 13: 0
CONSOLE MESSAGE: line 12: [object Object]
CONSOLE MESSAGE: line 13: [object Object]
CONSOLE MESSAGE: line 12: function Object() { [native code] }
CONSOLE MESSAGE: line 13: function Object() { [native code] }
CONSOLE MESSAGE: line 12: [object Object]
CONSOLE MESSAGE: line 13: [object Object]
CONSOLE MESSAGE: line 12: function ( /**/ foo/**/, /*/**/bar,
    /**/baz) {}
CONSOLE MESSAGE: line 13: function ( /**/ foo/**/, /*/**/bar,
    /**/baz) {}
CONSOLE MESSAGE: line 12: 42
CONSOLE MESSAGE: line 13: 42
CONSOLE MESSAGE: line 12: abc
CONSOLE MESSAGE: line 13: abc
CONSOLE MESSAGE: line 12: [object Uint16Array]
CONSOLE MESSAGE: line 13: [object Uint16Array]
CONSOLE MESSAGE: line 12: [object Text]
CONSOLE MESSAGE: line 13: [object Text]
CONSOLE MESSAGE: line 12: [object DOMException]
CONSOLE MESSAGE: line 13: [object DOMException]
CONSOLE MESSAGE: line 12: [object Uint8Array]
CONSOLE MESSAGE: line 13: [object Uint8Array]
CONSOLE MESSAGE: line 12: [object Uint8Array]
CONSOLE MESSAGE: line 13: [object Uint8Array]
Tests that console logging dumps proper messages.

 console-format.html:26 Array[10]
    0: "test"
    1: "test2"
    4: "test4"
    foo: Object
    length: 10
    __proto__: Array[0]
console-format.html:27 Array[10]
    0: "test"
    1: "test2"
    4: "test4"
    foo: Object
    length: 10
    __proto__: Array[0]
console-format.html:28 Array[10]
    0: "test"
    1: "test2"
    4: "test4"
    foo: Object
    length: 10
    __proto__: Array[0]
console-format.html:29 Test for zero "0" in formatter
console-format.html:30 % self-escape1 dummy
console-format.html:31 %s self-escape2 dummy
console-format.html:32 %ss self-escape3 dummy
console-format.html:33 %sdummy%s self-escape4
console-format.html:34 %%% self-escape5 dummy
console-format.html:35 %dummy self-escape6
console-format.html:12 /^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
console-format.html:13 [/^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\…?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i]
    0: /^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
    length: 1
    __proto__: Array[0]
globals[0]
/^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
console-format.html:12 /foo\\bar\sbaz/i
console-format.html:13 [/foo\\bar\sbaz/i]
    0: /foo\\bar\sbaz/i
    length: 1
    __proto__: Array[0]
globals[1]
/foo\\bar\sbaz/i
console-format.html:12 test
console-format.html:13 ["test"]
globals[2]
"test"
console-format.html:12 test named "test"
console-format.html:13 ["test named "test""]
globals[3]
"test named "test""
console-format.html:12 Error
console-format.html:13 [Error
]
    0: Error
    length: 1
    __proto__: Array[0]
globals[4]
Error
console-format.html:12 Error: my error message
console-format.html:13 [Error: my error message
]
    0: Error: my error message
    length: 1
    __proto__: Array[0]
globals[5]
Error: my error message
console-format.html:12 Error: my multiline
error message
console-format.html:13 [Error: my multiline
error message
]
    0: Error: my multiline
error message
    length: 1
    __proto__: Array[0]
globals[6]
Error: my multiline
error message
console-format.html:12 
    <p id="p">Tests that console logging dumps proper messages.</p>
console-format.html:13 [p#p]
    0: p#p
    length: 1
    __proto__: Array[0]
globals[7]
    <p id="p">Tests that console logging dumps proper messages.</p>
console-format.html:12 () { return 1; }
console-format.html:13 [function]
    0: ()
    length: 1
    __proto__: Array[0]
globals[8]
() { return 1; }
console-format.html:12 () {
        return 2;
    }
console-format.html:13 [function]
    0: ()
    length: 1
    __proto__: Array[0]
globals[9]
() {
        return 2;
    }
console-format.html:12 0.12
console-format.html:13 [0.12]
globals[10]
0.12
console-format.html:12 http://webkit.org/
console-format.html:13 ["http://webkit.org/"]
globals[11]
"http://webkit.org/"
console-format.html:12 null
console-format.html:13 [null]
globals[12]
null
console-format.html:12 undefined
console-format.html:13 [undefined]
globals[13]
undefined
console-format.html:12 
    attr=""
console-format.html:13 [attr]
    0: attr
    length: 1
    __proto__: Array[0]
globals[14]
    attr=""
console-format.html:12 
    attr="value"
console-format.html:13 [attr]
    0: attr
    length: 1
    __proto__: Array[0]
globals[15]
    attr="value"
console-format.html:12 
    id="x"
console-format.html:13 [id]
    0: id
    length: 1
    __proto__: Array[0]
globals[16]
    id="x"
console-format.html:12 Object {}
    length: (...)
    get length: length()
    __proto__: Object
console-format.html:13 [Object]
    0: Object
    length: 1
    __proto__: Array[0]
globals[17]
Object {}
    length: (...)
    get length: length()
    __proto__: Object
console-format.html:12 NaN
console-format.html:13 [NaN]
globals[18]
NaN
console-format.html:12 Infinity
console-format.html:13 [Infinity]
globals[19]
Infinity
console-format.html:12 -Infinity
console-format.html:13 [-Infinity]
globals[20]
-Infinity
console-format.html:12 ["test", "test2", 4: "test4", foo: Object]
    0: "test"
    1: "test2"
    4: "test4"
    foo: Object
    length: 10
    __proto__: Array[0]
console-format.html:13 [Array[10]]
    0: Array[10]
    length: 1
    __proto__: Array[0]
globals[21]
["test", "test2", undefined × 2, "test4", undefined × 5]
console-format.html:12 Object {}
console-format.html:13 [Object]
    0: Object
    length: 1
    __proto__: Array[0]
globals[22]
Object {}
console-format.html:12 [function]
    0: ()
    length: 1
    __proto__: Array[0]
console-format.html:13 [Array[1]]
    0: Array[1]
    length: 1
    __proto__: Array[0]
globals[23]
[anonymous()]
console-format.html:12 Object {bar: "bar"}
    bar: "bar"
    __proto__: Object
console-format.html:13 [Object]
    0: Object
    length: 1
    __proto__: Array[0]
globals[24]
Object {bar: "bar"}
    bar: "bar"
    __proto__: Object
console-format.html:12 
    <svg id="svg-node"></svg>
console-format.html:13 [svg#svg-node]
    0: svg#svg-node
    length: 1
    __proto__: Array[0]
globals[25]
    <svg id="svg-node"></svg>
console-format.html:12 Object {}
    bar: (...)
    get bar: ()
    set bar: (x)
    getFoo: ()
    __proto__: Object
console-format.html:13 [Object]
    0: Object
    length: 1
    __proto__: Array[0]
globals[26]
Object {}
    bar: (...)
    get bar: ()
    set bar: (x)
    getFoo: ()
    __proto__: Object
console-format.html:12 -0
console-format.html:13 [-0]
globals[27]
-0
console-format.html:12 Object {}
console-format.html:13 [Object]
    0: Object
    length: 1
    __proto__: Array[0]
globals[28]
Object {}
console-format.html:12 Object() { [native code] }
console-format.html:13 [function]
    0: Object()
    length: 1
    __proto__: Array[0]
globals[29]
Object() { [native code] }
console-format.html:12 Object {}
    __defineGetter__: __defineGetter__()
    __defineSetter__: __defineSetter__()
    __lookupGetter__: __lookupGetter__()
    __lookupSetter__: __lookupSetter__()
    constructor: Object()
    hasOwnProperty: hasOwnProperty()
    isPrototypeOf: isPrototypeOf()
    propertyIsEnumerable: propertyIsEnumerable()
    toLocaleString: toLocaleString()
    toString: toString()
    valueOf: valueOf()
    get __proto__: get __proto__()
    set __proto__: set __proto__()
console-format.html:13 [Object]
    0: Object
    length: 1
    __proto__: Array[0]
globals[30]
Object {}
    __defineGetter__: __defineGetter__()
    __defineSetter__: __defineSetter__()
    __lookupGetter__: __lookupGetter__()
    __lookupSetter__: __lookupSetter__()
    constructor: Object()
    hasOwnProperty: hasOwnProperty()
    isPrototypeOf: isPrototypeOf()
    propertyIsEnumerable: propertyIsEnumerable()
    toLocaleString: toLocaleString()
    toString: toString()
    valueOf: valueOf()
    get __proto__: get __proto__()
    set __proto__: set __proto__()
console-format.html:12 ( /**/ foo/**/, /*/**/bar,
    /**/baz) {}
console-format.html:13 [function]
    0: ( /**/ foo/**/, /*/**/bar,     /**/baz)
    length: 1
    __proto__: Array[0]
globals[31]
( /**/ foo/**/, /*/**/bar,
    /**/baz) {}
console-format.html:12 Number {[[PrimitiveValue]]: 42}
console-format.html:13 [Number]
    0: Number
    length: 1
    __proto__: Array[0]
globals[32]
Number {[[PrimitiveValue]]: 42}
console-format.html:12 String {0: "a", 1: "b", 2: "c", length: 3, [[PrimitiveValue]]: "abc"}
console-format.html:13 [String]
    0: String
    length: 1
    __proto__: Array[0]
globals[33]
String {0: "a", 1: "b", 2: "c", length: 3, [[PrimitiveValue]]: "abc"}
console-format.html:12 [1, 2, 3]
console-format.html:13 [Uint16Array[3]]
    0: Uint16Array[3]
    length: 1
    __proto__: Array[0]
globals[34]
[1, 2, 3]
console-format.html:12 #text
console-format.html:13 [text]
    0: text
    length: 1
    __proto__: Array[0]
globals[35]
#text
console-format.html:12 DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
console-format.html:13 [DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of th…]
    0: DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
    length: 1
    __proto__: Array[0]
globals[36]
DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
console-format.html:12 [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…]
    [0 … 99]
    [100 … 199]
    [200 … 299]
    [300 … 399]
    foo: "bar"
    __proto__: Uint8Array
console-format.html:13 [Uint8Array[400]]
    0: Uint8Array[400]
    length: 1
    __proto__: Array[0]
globals[37]
Uint8Array[400]
    [0 … 99]
    [100 … 199]
    [200 … 299]
    [300 … 399]
    foo: "bar"
    __proto__: Uint8Array
console-format.html:12 [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…]
    [0 … 99999999]
    [100000000 … 199999999]
    [200000000 … 299999999]
    [300000000 … 399999999]
    __proto__: Uint8Array
console-format.html:13 [Uint8Array[400000000]]
    0: Uint8Array[400000000]
    length: 1
    __proto__: Array[0]
globals[38]
Uint8Array[400000000]
    [0 … 99999999]
    [100000000 … 199999999]
    [200000000 … 299999999]
    [300000000 … 399999999]
    __proto__: Uint8Array
